+++ src/PoolAllocator.h
@@ -1,25 +1,26 @@
 #pragma once
 #include "mytypes.h"
+#include <stddef.h> // for size_t
 
 class PoolAllocator
 {
     struct block {
-        u64 free_size;
+        size_t free_size;
         struct block *next;
         u8 *start_address;
         u8 *free_address;
     };
     block root_block;
-    u64 total_size;
-    u64 block_size;
+    size_t total_size;
+    size_t block_size;
 
     void allocateBlock(block *b);
-    void * allocateFromBlock(block *b, u64 size);
+    void * allocateFromBlock(block *b, size_t size);
 public:
-    PoolAllocator(u64 start_size = 0);
+    PoolAllocator(size_t start_size = 0);
     ~PoolAllocator();
-    void * alloc(u64 size);
+    void * alloc(size_t size);
     bool isAddressInRange(void *p);
 };
 
-void * operator new (u64 size, PoolAllocator *p);
+void * operator new ( size_t size, PoolAllocator *p);
